correcting fc

+++ lib/features/feature_weather/data/models/for_cast_city_model.dart
@@ -12,20 +12,20 @@ import 'package:wheather_app/features/feature_weather/domain/entites/forecase_da
 
 class ForecastDaysModel extends ForecastDaysEntity{
   ForecastDaysModel({
-      double? lat, 
-      double? lon, 
-      String? timezone,
-      int? timezoneOffset, 
-      Current? current, 
-      List<Daily>? daily, 
-      List<Alerts>? alerts,}) : super(
-    lat: lat,
-    lon: lon,
-    timezone: timezone,
-    timezoneOffset: timezoneOffset,
-    current: current,
-    daily: daily,
-    alerts: alerts
+    double? lat,
+    double? lon,
+    String? timezone,
+    int? timezoneOffset,
+    Current? current,
+    List<Daily>? daily,
+    List<Alerts>? alerts,}) : super(
+      lat: lat,
+      lon: lon,
+      timezone: timezone,
+      timezoneOffset: timezoneOffset,
+      current: current,
+      daily: daily,
+      alerts: alerts
   );
 
   factory ForecastDaysModel.fromJson(dynamic json) {
@@ -47,13 +47,13 @@ class ForecastDaysModel extends ForecastDaysEntity{
     }
 
     return ForecastDaysModel(
-      lat: json['lat'],
-      lon: json['lon'],
-      timezone: json['timezone'],
-      timezoneOffset: json['timezone_offset'],
-      current: json['current'] != null ? Current.fromJson(json['current']) : null,
-      daily: daily,
-      alerts: alerts
+        lat: json['lat'].toDouble(),
+        lon: json['lon'].toDouble(),
+        timezone: json['timezone'],
+        timezoneOffset: json['timezone_offset'],
+        current: json['current'] != null ? Current.fromJson(json['current']) : null,
+        daily: daily,
+        alerts: alerts
     );
 
 
@@ -70,19 +70,19 @@ class ForecastDaysModel extends ForecastDaysEntity{
 
 class Alerts {
   Alerts({
-      String? senderName, 
-      String? event, 
-      int? start, 
-      int? end, 
-      String? description, 
-      List<String>? tags,}){
+    String? senderName,
+    String? event,
+    int? start,
+    int? end,
+    String? description,
+    List<String>? tags,}){
     _senderName = senderName;
     _event = event;
     _start = start;
     _end = end;
     _description = description;
     _tags = tags;
-}
+  }
 
   Alerts.fromJson(dynamic json) {
     _senderName = json['sender_name'];
@@ -141,25 +141,25 @@ class Alerts {
 
 class Daily {
   Daily({
-      int? dt,
-      int? sunrise,
-      int? sunset,
-      int? moonrise, 
-      int? moonset, 
-      double? moonPhase, 
-      Temp? temp, 
-      Feels_like? feelsLike, 
-      int? pressure, 
-      int? humidity, 
-      double? dewPoint, 
-      double? windSpeed, 
-      int? windDeg, 
-      double? windGust, 
-      List<Weather>? weather, 
-      int? clouds, 
-      double? pop, 
-      double? rain, 
-      double? uvi,}){
+    int? dt,
+    int? sunrise,
+    int? sunset,
+    int? moonrise,
+    int? moonset,
+    double? moonPhase,
+    Temp? temp,
+    Feels_like? feelsLike,
+    int? pressure,
+    int? humidity,
+    double? dewPoint,
+    double? windSpeed,
+    int? windDeg,
+    double? windGust,
+    List<Weather>? weather,
+    int? clouds,
+    double? pop,
+    double? rain,
+    double? uvi,}){
     _dt = dt;
     _sunrise = sunrise;
     _sunset = sunset;
@@ -179,7 +179,7 @@ class Daily {
     _pop = pop;
     _rain = rain;
     _uvi = uvi;
-}
+  }
 
   Daily.fromJson(dynamic json) {
     _dt = json['dt'];
@@ -289,15 +289,15 @@ class Daily {
 
 class Weather {
   Weather({
-      int? id, 
-      String? main, 
-      String? description, 
-      String? icon,}){
+    int? id,
+    String? main,
+    String? description,
+    String? icon,}){
     _id = id;
     _main = main;
     _description = description;
     _icon = icon;
-}
+  }
 
   Weather.fromJson(dynamic json) {
     _id = json['id'];
@@ -333,15 +333,15 @@ class Weather {
 
 class Feels_like {
   Feels_like({
-      double? day, 
-      double? night, 
-      double? eve, 
-      double? morn,}){
+    double? day,
+    double? night,
+    double? eve,
+    double? morn,}){
     _day = day;
     _night = night;
     _eve = eve;
     _morn = morn;
-}
+  }
 
   Feels_like.fromJson(dynamic json) {
     _day = json['day'].toDouble();
@@ -379,19 +379,19 @@ class Feels_like {
 
 class Temp {
   Temp({
-      double? day, 
-      double? min, 
-      double? max, 
-      double? night, 
-      double? eve, 
-      double? morn,}){
+    double? day,
+    double? min,
+    double? max,
+    double? night,
+    double? eve,
+    double? morn,}){
     _day = day;
     _min = min;
     _max = max;
     _night = night;
     _eve = eve;
     _morn = morn;
-}
+  }
 
   Temp.fromJson(dynamic json) {
     _day = json['day'].toDouble();
@@ -446,21 +446,21 @@ class Temp {
 
 class Current {
   Current({
-      int? dt,
-      int? sunrise,
-      int? sunset,
-      double? temp, 
-      double? feelsLike, 
-      int? pressure, 
-      int? humidity, 
-      double? dewPoint, 
-      double? uvi, 
-      int? clouds, 
-      int? visibility, 
-      double? windSpeed, 
-      int? windDeg, 
-      double? windGust, 
-      List<Weather>? weather,}){
+    int? dt,
+    int? sunrise,
+    int? sunset,
+    double? temp,
+    double? feelsLike,
+    int? pressure,
+    int? humidity,
+    double? dewPoint,
+    double? uvi,
+    int? clouds,
+    int? visibility,
+    double? windSpeed,
+    int? windDeg,
+    double? windGust,
+    List<Weather>? weather,}){
     _dt = dt;
     _sunrise = sunrise;
     _sunset = sunset;
@@ -476,7 +476,7 @@ class Current {
     _windDeg = windDeg;
     _windGust = windGust;
     _weather = weather;
-}
+  }
 
   Current.fromJson(dynamic json) {
     _dt = json['dt'];
